Fixup build of the main binary and get rid of linking errors

+++ generator/date_range.cpp
@@ -4,7 +4,7 @@ using namespace std::chrono;
 generator::DateRange::DateRange(const sys_days minValue, const sys_days maxValue)
   : minValue_(minValue)
   , maxValue_(maxValue)
-  , distribution_(std::uniform_int_distribution((maxValue - minValue).count())) {
+  , distribution_(std::uniform_int_distribution(static_cast<int64_t>((maxValue - minValue).count()))) {
 }
 
 sys_days generator::DateRange::next() {
